chore(kit-checklist): refactoring code for readability
